+++ src/common/tetris_lib/tetromino/tetromino.cpp
@@ -59,8 +59,8 @@ ATetromino::ATetromino(Vec2 &&anchorPoint, std::vector<Vec2> &&body,
         maxX = std::max(maxX, coord.getX());
     }
 
-    height_ = maxY - minY + 1;
-    width_ = maxX - minX + 1;
+    height_ = static_cast<size_t>(maxY - minY + 1);
+    width_ = static_cast<size_t>(maxX - minX + 1);
 }
 
 /*--------------------------------------------------
@@ -139,8 +139,8 @@ uint8_t ATetromino::getNumOfTests() const noexcept {
 TetrominoPtr ATetromino::getNthOffset(uint8_t offsetIndex) const {
     TetrominoPtr copy = this->clone();
 
-    Vec2 offsetVal1 = offsetData_->at(prevRotationIdx_).at(offsetIndex - 1);
-    Vec2 offsetVal2 = offsetData_->at(rotationIdx_).at(offsetIndex - 1);
+    Vec2 offsetVal1 = offsetData_->at(static_cast<size_t>(prevRotationIdx_)).at(offsetIndex - 1);
+    Vec2 offsetVal2 = offsetData_->at(static_cast<size_t>(rotationIdx_)).at(offsetIndex - 1);
 
     // Compute offset with offset data
     Vec2 offset = (offsetVal1 - offsetVal2);
